Round down deterministic-K input at the edge of the group size, not above it.
I believe the implementation has a tiny mistake that occurs when the input happens
to equal the group order (eg z = p256k1_N = 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFEBAAEDCE6AF48A03BBFD25E8CD0364141).

https://tools.ietf.org/html/rfc6979#section-2.3.4 says:
>    2.  z1 is reduced modulo q, yielding z2 (an integer between 0 and
>       q-1, inclusive):
>
>          z2 = z1 mod q
>
>       Note that since z1 is less than 2^qlen, that modular reduction
>       can be implemented with a simple conditional subtraction:
>       z2 = z1-q if that value is non-negative; otherwise, z2 = z1.

Here val = z1, and n = q. It's saying:
```
z2 = z1 - q if (z1 - q >= 0) else z1
```
which is the same as
val = z1; n = q;
if val - n >= 0:
  val = val - n

which is the same as
if val >= n:
  val -= n

Not doing this makes this case come out to the wrong hash, even though all other inputs are correct.